+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-## [0.1.2] - 2024-11-19
+## [0.2.0] - 2024-11-19
 ### Added
 - Support for wasm32-unknown-unknown target.
-- Implement `TrafficShaperPayload` for `Result` and `Option`.
+- Implement `ChokeItem` for `Result` and `Option`.
+- Rename types.